split cost out into its own module
Added a function to insert a new cost into a list, which could be used to
asjust costs after a drag and drop.

+++ Assistant/Types/ScanRemotes.hs
@@ -13,7 +13,7 @@ import Control.Concurrent.STM
 import qualified Data.Map as M
 
 data ScanInfo = ScanInfo
-	{ scanPriority :: Int
+	{ scanPriority :: Float
 	, fullScan :: Bool
 	}
